requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
bidict==0.22.1
boto3==1.28.17
botocore==1.31.17
CacheControl==0.12.14
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==2.2.0
cryptography==41.0.3
cycler==0.11.0
dask==2022.10.2
dnspython==2.4.1
docopt==0.6.2
fbmessenger==6.0.0
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
jax==0.4.14
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.1
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.6
locket==1.0.0
magic-filter==1.0.11
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
mattermostwrapper==2.2
ml-dtypes==0.2.0
msgpack==1.0.5
multidict==5.2.0
networkx==2.6.3
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
pandas==2.0.3
partd==1.4.0
Pillow==10.0.0
pluggy==1.2.0
portalocker==2.7.0
prompt-toolkit==3.0.28
protobuf==4.23.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pydot==1.4.2
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.1.1
pyrsistent==0.19.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2022.7.1
PyYAML==6.0.1
questionary==1.10.0
randomname==0.1.5
rasa==3.6.4
rasa-sdk==3.6.1
redis==4.6.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rocketchat-API==1.30.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.8.1
seaborn==0.12.2
sentry-sdk==1.14.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.21.3
sns==0.1
SQLAlchemy==1.4.49
structlog==23.1.0
structlog-sentry==2.0.3
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-macos==2.12.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.2.0
toolz==0.12.0
tqdm==4.65.0
twilio==8.2.2
typing-extensions==4.7.1
typing-utils==0.1.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==1.26.16
uvloop==0.17.0
wcwidth==0.2.6
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.3.6
wrapt==1.14.1
yarl==1.9.2